+++ ricaun.Revit.Mvvm/Input/AsyncRelayCommand.cs
@@ -30,7 +30,7 @@ namespace ricaun.Revit.Mvvm
         /// <summary>
         /// The optional action to Handle Exception
         /// </summary>
-        private Action<Exception> handleException;
+        public event Action<Exception> ExceptionHandler;
 
         /// <summary>
         /// Set Exception Handler 
@@ -39,7 +39,7 @@ namespace ricaun.Revit.Mvvm
         /// <returns></returns>
         public AsyncRelayCommand SetExceptionHandler(Action<Exception> handleException)
         {
-            this.handleException = handleException;
+            this.ExceptionHandler = handleException;
             return this;
         }
 
@@ -99,7 +99,7 @@ namespace ricaun.Revit.Mvvm
         /// <param name="parameter"></param>
         public void Execute(object parameter)
         {
-            this.ExecuteAsync(parameter).Run(handleException);
+            this.ExecuteAsync(parameter).Run(ExceptionHandler);
         }
 
         /// <summary>