+++ ricaun.Revit.Mvvm/Input/RelayCommand{T}.cs
@@ -28,7 +28,7 @@ namespace ricaun.Revit.Mvvm
         /// <summary>
         /// The optional action to Handle Exception
         /// </summary>
-        private Action<Exception> handleException;
+        public event Action<Exception> ExceptionHandler;
 
         /// <summary>
         /// Set Exception Handler 
@@ -37,7 +37,7 @@ namespace ricaun.Revit.Mvvm
         /// <returns></returns>
         public RelayCommand<T> SetExceptionHandler(Action<Exception> handleException)
         {
-            this.handleException = handleException;
+            this.ExceptionHandler = handleException;
             return this;
         }
 
@@ -113,7 +113,7 @@ namespace ricaun.Revit.Mvvm
                 }
                 catch (Exception ex)
                 {
-                    handleException?.Invoke(ex);
+                    ExceptionHandler?.Invoke(ex);
                 }
             }
         }